feat(user_profile): add golden tests and refactor widgets
- Extract ProfileHeader, ProfileSection, and ProfileInfoTile to separate widgets for better testability.
- Update UserProfilePage to use the new public widgets.
- Add comprehensive golden tests for UserProfilePage (Loaded, Loading, Error states).
- Add isolated golden tests for ProfileHeader and ProfileSection.
- Generate golden reference images in test/features/user_profile/presentation/golden/goldens/.
- Remove redundant golden test from presentation/pages/.
